refactor : make useResolveMenuBar
