+++ src/hooks/usePreview.ts
@@ -1,4 +1,4 @@
-import React, { useState, useEffect, useMemo } from 'react';
+import { useState, useEffect, useMemo } from 'react';
 
 export const usePreview = () => {
   const [file, setFile] = useState<File>();
